# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockTank API",
    "version": "0.1.0",
    "description": "Identity, RBAC, public content, advertising, marketing and system endpoints for StockTank. Sessions are cookie-based; state-changing requests must send `X-Requested-With`. Financial content served by this platform is informational only and never investment advice."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (dev proxy or deployed API host)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "st_session",
        "description": "httpOnly session cookie set by /api/v1/auth/register and /api/v1/auth/login. The server stores only a SHA-256 hash of the token."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Envelope used by every non-2xx response.",
        "example": {
          "error": {
            "code": "VALIDATION_FAILED",
            "message": "Invalid request body",
            "requestId": "7f1c1b7e-2d1a-4a4d-9d2b-3b7b2a1f0c9e",
            "details": {
              "source": "body",
              "issues": [
                {
                  "path": "email",
                  "message": "Invalid email address",
                  "code": "invalid_format"
                }
              ]
            }
          }
        }
      },
...
      "ShowListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "tagline": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "coverUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "episodeCount": {
                  "type": "integer"
                },
                "isDemo": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "slug",
                "title",
                "tagline",
                "description",
                "coverUrl",
                "episodeCount",
                "isDemo"
              ]
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "ShowDetailResponse": {
        "type": "object",
        "properties": {
          "show": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "tagline": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "description": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "coverUrl": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "episodeCount": {
                "type": "integer"
              },
              "isDemo": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "slug",
              "title",
              "tagline",
              "description",
              "coverUrl",
              "episodeCount",
              "isDemo"
            ]
          },
          "episodes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "summary": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "coverUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "durationSeconds": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "publishedAt": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "show": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "slug",
                    "title"
                  ]
                },
                "isDemo": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "slug",
                "title",
                "summary",
                "coverUrl",
                "durationSeconds",
                "publishedAt",
                "show",
                "isDemo"
              ]
            }
          }
        },
        "required": [
          "show",
          "episodes"
        ]
      },
      "ProjectListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "symbol": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "kind": {
                  "type": "string",
                  "enum": [
                    "crypto_project",
                    "protocol",
                    "dao",
                    "infrastructure",
                    "rwa",
                    "ecosystem",
                    "application",
                    "traditional_company"
                  ]
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "chainName": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "logoUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "verified": {
                  "type": "boolean"
                },
                "isDemo": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "slug",
                "name",
                "symbol",
                "kind",
                "description",
                "chainName",
                "logoUrl",
                "verified",
                "isDemo"
              ]
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "CompanyListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "ticker": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "exchange": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "sector": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "country": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "logoUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "isDemo": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "slug",
                "name",
                "ticker",
                "exchange",
                "sector",
                "country",
                "description",
                "logoUrl",
                "isDemo"
              ]
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
...
      "PublicFlagsResponse": {
        "type": "object",
        "properties": {
          "flags": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "flags"
        ],
        "example": {
          "flags": {
            "advertising": false,
            "wallet": false
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/home": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "Homepage",
        "description": "Featured shows, latest episodes, published clips, explainers, projects, companies, the live broadcast and today’s rundown. Cached 30s.",
        "responses": {
          "200": {
            "description": "Homepage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shows": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "List shows",
        "description": "Published shows, alphabetical.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 24
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List shows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shows/{slug}": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "Show detail",
        "description": "A published show and its published episodes.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Show detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Show not found (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "List projects",
        "description": "Published project profiles. Informational only.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 24
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List projects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "List companies",
        "description": "Published company profiles. Informational only.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 24
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List companies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "Search",
        "description": "Grouped search across shows, episodes, projects and companies (Postgres until Meilisearch).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "required": true,
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Search",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/flags": {
      "get": {
        "tags": [
          "Content"
        ],
        "summary": "Feature flags",
        "description": "Public on/off state of every feature flag (§50).",
        "responses": {
          "200": {
            "description": "Feature flags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicFlagsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}